requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
click
itsdangerous
Jinja2
MarkupSafe
python-dotenv
Werkzeug
numpy
pandas
requests
flask_cors
geopandas
matplotlib
shapely
DBSCAN
metrics
StandardScaler
ortools
scikit-learn